backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
docker==6.0.1
python-dotenv==1.0.0
